Fdb: fix GetInt64(...) in FdbSystemStatus that was overflowing on 32 bits instead of 64-bits
- Would throw OverflowException when trying to read, for example, Memory.FreeBytes or equivalent that was overflowing a 32 bit int

+++ FoundationDB.Client/Status/FdbSystemStatus.cs
@@ -264,7 +264,7 @@ namespace FoundationDB.Client.Status
 
 		protected int? GetInt32(string field1, string field2) => m_data != null && m_data.TryGetObject(field1, out var v1) && v1.TryGet<int>(field2, out var v2) ? v2 : null;
 
-		protected long? GetInt64(string field) => m_data?.Get<int?>(field, null);
+		protected long? GetInt64(string field) => m_data?.Get<long?>(field, null);
 
 		protected long? GetInt64(string field1, string field2) => m_data != null && m_data.TryGetObject(field1, out var v1) && v1.TryGet<long>(field2, out var v2) ? v2 : null;
